Fixes, removal of block-global-framework param
