+++ source/class/qx/tool/cli/commands/Lint.js
@@ -16,10 +16,9 @@
 
 ************************************************************************ */
 
-
-
 const CLIEngine = require("eslint").CLIEngine;
 const fs = qx.tool.utils.Promisify.fs;
+const replaceInFile = require("replace-in-file");
 
 qx.Class.define("qx.tool.cli.commands.Lint", {
   extend: qx.tool.cli.commands.Command,
@@ -174,7 +173,7 @@ qx.Class.define("qx.tool.cli.commands.Lint", {
           to: "@param $2 $1"
         });
       }
-      await qx.tool.migration.Utils.replaceInFiles(replaceInFiles);
+      await replaceInFile(replaceInFiles);
     }
   }
 });
